+++ src/components/Chat/ChatMessage.tsx
@@ -2,8 +2,21 @@ import * as React from 'react'
 import * as PropTypes from 'prop-types'
 
 import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
+import { ComponentVariablesInput, IComponentPartStylesInput } from '../../../types/theme'
+import { Extendable, ReactChildren } from '../../../types/utils'
+import Avatar, { IAvatarProps } from '../Avatar/Avatar'
+export interface IChatMessageProps {
+  as?: any
+  avatar?: IAvatarProps
+  children?: ReactChildren
+  className?: string
+  content?: any
+  mine?: boolean
+  styles?: IComponentPartStylesInput
+  variables?: ComponentVariablesInput
+}
 
-class ChatMessage extends UIComponent<any, any> {
+class ChatMessage extends UIComponent<Extendable<IChatMessageProps>, any> {
   static className = 'ui-chat__message'
 
   static create: Function
@@ -12,6 +25,9 @@ class ChatMessage extends UIComponent<any, any> {
 
   static propTypes = {
     as: customPropTypes.as,
+
+    /** Avatar */
+    avatar: PropTypes.object,
 
     /** Child content. */
     children: PropTypes.node,
@@ -32,21 +48,37 @@ class ChatMessage extends UIComponent<any, any> {
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
-  static handledProps = ['as', 'children', 'className', 'content', 'mine', 'styles', 'variables']
+  static handledProps = [
+    'as',
+    'avatar',
+    'children',
+    'className',
+    'content',
+    'mine',
+    'styles',
+    'variables',
+  ]
 
   static defaultProps = {
     as: 'li',
   }
 
-  renderComponent({ ElementType, classes, rest }) {
-    const { children, content } = this.props
+  renderComponent({ ElementType, classes, rest, styles }) {
+    const { avatar, children, content, mine } = this.props
 
     return (
       <ElementType {...rest} className={classes.root}>
-        {childrenExist(children) ? children : content}
+        {!mine && this.renderAvatar(avatar, styles)}
+        <div className={classes.chatContent}>{childrenExist(children) ? children : content}</div>
+        {mine && this.renderAvatar(avatar, styles)}
       </ElementType>
     )
   }
+
+  private renderAvatar = (
+    avatarProps: IAvatarProps | undefined,
+    styles: IComponentPartStylesInput,
+  ) => avatarProps && <Avatar {...avatarProps} styles={{ root: styles.chatAvatar }} />
 }
 
 ChatMessage.create = createShorthandFactory(ChatMessage, content => ({ content }))